+++ mapfishapp/src/main/webapp/app/js/GEOR_Lang/ru.js
@@ -153,6 +153,9 @@ OpenLayers.Lang.ru = OpenLayers.Util.extend(OpenLayers.Lang.ru, {
     "Unreachable server or insufficient rights": "Ответ сервера недействителен. Возможные причины : недостаточные права, сервер недостижим, слишком много данных, и т.д...",
     "WFS server": "Сервер WFS",
     /* GEOR_managelayers.js strings */
+    //"Service": "Service",
+    //"Protocol": "Protocole",
+    //"About this layer": "A propos de cette couche",
     //"Set as overlay": "Passer en calque",
     //"Set as baselayer": "Passer en couche de fond",
     "Confirm NAME layer deletion ?": "Вы действительно хотите удалить слой${NAME} ?",